Refactor dashboard workflows
