+++ src/components/dashboard/HybridGradeImportDialog.tsx
@@ -1,9 +1,5 @@
-import { useEffect, useMemo, useState, type ChangeEvent } from "react";
 import type { ComponentProps, ReactNode } from "react";
 import { ArrowRight, CheckCircle2, FileSpreadsheet, Image as ImageIcon, Loader2, Upload } from "lucide-react";
-import { toast } from "sonner";
-
-import { supabase } from "@/integrations/supabase/client";
 import { Button } from "@/components/ui/button";
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
@@ -13,11 +9,11 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Badge } from "@/components/ui/badge";
 import { Card, CardContent } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
 import { Textarea } from "@/components/ui/textarea";
-import { cn } from "@/lib/utils";
 import { getHybridGradeImportTemplateHref, isHybridGradeImportEnabled } from "@/lib/hybridImport";
 import { safeFormatDate } from "@/lib/date";
+import { HybridGradeImportPreview } from "@/components/dashboard/hybrid-grade-import/HybridGradeImportPreview";
+import { useHybridGradeImportDialog } from "@/components/dashboard/hybrid-grade-import/useHybridGradeImportDialog";
 
 type AssignmentOption = {
   id: string;
@@ -124,330 +120,42 @@ export const HybridGradeImportDialog = ({
   triggerClassName?: string;
   defaultMode?: "csv" | "image";
 }) => {
-  const [open, setOpen] = useState(false);
-  const [mode, setMode] = useState<"csv" | "image">(defaultMode);
-  const [importScope, setImportScope] = useState<ImportScope>(assignments.length > 0 ? "existing_assignment" : "new_assignment");
-  const [assignmentId, setAssignmentId] = useState("");
-  const [draft, setDraft] = useState<DraftState>({ csvText: "", csvFileName: "", imageFiles: [] });
-  const [newAssignment, setNewAssignment] = useState<NewAssignmentDraft>({
-    title: `Imported grades - ${new Date().toISOString().slice(0, 10)}`,
-    moduleCode: "",
-    maxScore: "100",
-    dueDate: "",
-    description: "",
-  });
-  const [preview, setPreview] = useState<GradeImportResponse | null>(null);
-  const [committed, setCommitted] = useState<GradeImportResponse | null>(null);
-  const [loadingPreview, setLoadingPreview] = useState(false);
-  const [loadingCommit, setLoadingCommit] = useState(false);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
-
-  const selectedAssignment = useMemo(
-    () => assignments.find((assignment) => assignment.id === assignmentId) ?? assignments[0] ?? null,
-    [assignmentId, assignments],
-  );
-
-  useEffect(() => {
-    if (!open) return;
-    if (importScope === "existing_assignment" && !assignmentId && assignments.length > 0) {
-      setAssignmentId(assignments[0].id);
-    }
-  }, [assignmentId, assignments, importScope, open]);
-
-  useEffect(() => {
-    if (!open) {
-      setMode(defaultMode);
-      setImportScope(assignments.length > 0 ? "existing_assignment" : "new_assignment");
-      setAssignmentId(assignments[0]?.id ?? "");
-      setDraft({ csvText: "", csvFileName: "", imageFiles: [] });
-      setNewAssignment({
-        title: `Imported grades - ${new Date().toISOString().slice(0, 10)}`,
-        moduleCode: "",
-        maxScore: "100",
-        dueDate: "",
-        description: "",
-      });
-      setPreview(null);
-      setCommitted(null);
-      setErrorMessage(null);
-    }
-  }, [assignments, defaultMode, open]);
+  const controller = useHybridGradeImportDialog(assignments, defaultMode);
+  const {
+    open,
+    setOpen,
+    mode,
+    setMode,
+    importScope,
+    setImportScope,
+    assignmentId,
+    setAssignmentId,
+    draft,
+    setDraft,
+    newAssignment,
+    setNewAssignment,
+    selectedAssignment,
+    preview,
+    setPreview,
+    committed,
+    setCommitted,
+    loadingPreview,
+    loadingCommit,
+    errorMessage,
+    setErrorMessage,
+    canPreview,
+    hasAcceptedRows,
+    confirmLabel,
+    usingExistingAssignment,
+    handleCsvFileChange,
+    handleImageFileChange,
+    handlePreview,
+    handleCommit,
+  } = controller;
 
   if (!isHybridGradeImportEnabled()) {
     return null;
   }
-
-  const activeFiles = mode === "csv" ? [] : draft.imageFiles;
-  const usingExistingAssignment = importScope === "existing_assignment";
-  const selectedNewAssignmentMaxScore = Number(newAssignment.maxScore) || 100;
-  const canPreview = usingExistingAssignment
-    ? Boolean(selectedAssignment) && (mode === "csv" ? draft.csvText.trim().length > 0 : activeFiles.length > 0)
-    : Boolean(newAssignment.title.trim()) &&
-      Boolean(newAssignment.moduleCode.trim()) &&
-      Number.isFinite(selectedNewAssignmentMaxScore) &&
-      selectedNewAssignmentMaxScore > 0 &&
-      (mode === "csv" ? draft.csvText.trim().length > 0 : activeFiles.length > 0);
-  const hasAcceptedRows = Boolean(preview?.rows.some((row) => row.accepted));
-  const confirmLabel = preview
-    ? `Import ${preview.summary.rowsAccepted} grade${preview.summary.rowsAccepted === 1 ? "" : "s"}`
-    : "Import grades";
-
-  const buildBody = (confirm: boolean) => {
-    if (mode === "csv") {
-      if (!draft.csvText.trim()) {
-        throw new Error("Paste or upload CSV content first.");
-      }
-      const baseBody: Record<string, unknown> = {
-        confirm,
-        importMethod: "csv" as const,
-        createMissingSubmissions: true,
-        csvText: draft.csvText.trim(),
-        sourceFileName: draft.csvFileName || "grades.csv",
-        importScope,
-      };
-
-      if (usingExistingAssignment) {
-        if (!selectedAssignment) {
-          throw new Error("Select an assignment before importing.");
-        }
-        return {
-          ...baseBody,
-          assignmentId: selectedAssignment.id,
-        };
-      }
-
-      return {
-        ...baseBody,
-        newAssignmentTitle: newAssignment.title.trim(),
-        newAssignmentModuleCode: newAssignment.moduleCode.trim(),
-        newAssignmentMaxScore: selectedNewAssignmentMaxScore,
-        newAssignmentDueDate: newAssignment.dueDate || null,
-        newAssignmentDescription: newAssignment.description.trim() || null,
-      };
-    }
-
-    if (draft.imageFiles.length === 0) {
-      throw new Error("Upload one or more images first.");
-    }
-
-    const formData = new FormData();
-    formData.append("confirm", String(confirm));
-    formData.append("importMethod", "image");
-    formData.append("createMissingSubmissions", "true");
-    formData.append("importScope", importScope);
-    if (usingExistingAssignment) {
-      if (!selectedAssignment) {
-        throw new Error("Select an assignment before importing.");
-      }
-      formData.append("assignmentId", selectedAssignment.id);
-    } else {
-      formData.append("newAssignmentTitle", newAssignment.title.trim());
-      formData.append("newAssignmentModuleCode", newAssignment.moduleCode.trim());
-      formData.append("newAssignmentMaxScore", String(selectedNewAssignmentMaxScore));
-      if (newAssignment.dueDate) formData.append("newAssignmentDueDate", newAssignment.dueDate);
-      if (newAssignment.description.trim()) formData.append("newAssignmentDescription", newAssignment.description.trim());
-    }
-    for (const file of draft.imageFiles) {
-      formData.append("file", file, file.name);
-    }
-    return formData;
-  };
-
-  const invokeImport = async (confirm: boolean) => {
-    const { data, error } = await supabase.functions.invoke("import-grades", {
-      body: buildBody(confirm),
-    });
-    if (error) {
-      throw error;
-    }
-    return data as GradeImportResponse;
-  };
-
-  const readFileText = async (file: File) => file.text();
-
-  const handleCsvFileChange = async (event: ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (!file) return;
-    setDraft({
-      csvText: await readFileText(file),
-      csvFileName: file.name,
-      imageFiles: [],
-    });
-    setMode("csv");
-    setPreview(null);
-    setCommitted(null);
-    setErrorMessage(null);
-    event.target.value = "";
-  };
-
-  const handleImageFileChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const files = Array.from(event.target.files ?? []);
-    if (files.length === 0) return;
-    setDraft({
-      csvText: "",
-      csvFileName: "",
-      imageFiles: files,
-    });
-    setMode("image");
-    setPreview(null);
-    setCommitted(null);
-    setErrorMessage(null);
-    event.target.value = "";
-  };
-
-  const handlePreview = async () => {
-    if (usingExistingAssignment && !selectedAssignment) {
-      toast.error("Choose an assignment first.");
-      return;
-    }
-
-    if (!canPreview) {
-      if (usingExistingAssignment) {
-        toast.error(mode === "csv" ? "Add CSV content first." : "Upload image files first.");
-      } else {
-        toast.error("Fill in the new assignment details and add the CSV or image first.");
-      }
-      return;
-    }
-
-    setLoadingPreview(true);
-    setErrorMessage(null);
-    try {
-      const data = await invokeImport(false);
-      if (!data || typeof data !== "object" || !Array.isArray((data as GradeImportResponse).rows)) {
-        throw new Error("Import preview returned an unexpected response.");
-      }
-
-      setPreview(data as GradeImportResponse);
-      setCommitted(null);
-    } catch (error) {
-      const message = error instanceof Error ? error.message : "Import preview failed";
-      setErrorMessage(message);
-      toast.error(message);
-    } finally {
-      setLoadingPreview(false);
-    }
-  };
-
-  const handleCommit = async () => {
-    if (!preview) {
-      toast.error("Preview the import before confirming it.");
-      return;
-    }
-
-    if (!hasAcceptedRows) {
-      toast.error("There are no valid rows to import.");
-      return;
-    }
-
-    setLoadingCommit(true);
-    setErrorMessage(null);
-    try {
-      const data = await invokeImport(true);
-      if (!data || typeof data !== "object" || (data as GradeImportResponse).committed !== true) {
-        throw new Error("Import confirmation returned an unexpected response.");
-      }
-
-      setCommitted(data as GradeImportResponse);
-      setPreview(data as GradeImportResponse);
-      toast.success("Grade import completed.");
-    } catch (error) {
-      const message = error instanceof Error ? error.message : "Grade import failed";
-      setErrorMessage(message);
-      toast.error(message);
-    } finally {
-      setLoadingCommit(false);
-    }
-  };
-
-  const renderPreview = () => {
-    if (!preview) {
-      return (
-        <div className="rounded-xl border border-dashed bg-muted/20 p-6 text-sm text-muted-foreground">
-          {mode === "csv"
-            ? "Upload or paste a CSV, then preview the import."
-            : "Upload image files, then preview the extracted rows."}
-        </div>
-      );
-    }
-
-    return (
-      <div className="space-y-3">
-        <div className="flex flex-wrap items-center gap-2">
-          <Badge variant="outline">{preview.summary.rowsProcessed} rows scanned</Badge>
-          <Badge variant="outline" className="border-success/30 bg-success/5 text-success">
-            {preview.summary.rowsAccepted} accepted
-          </Badge>
-          <Badge variant="outline" className="border-warning/30 bg-warning/5 text-warning">
-            {preview.summary.rowsWithWarnings} with warnings
-          </Badge>
-          <Badge variant="outline" className="border-destructive/30 bg-destructive/5 text-destructive">
-            {preview.summary.rowsRejected} rejected
-          </Badge>
-        </div>
-
-        <div className="overflow-hidden rounded-xl border">
-          <Table>
-            <TableHeader>
-              <TableRow>
-                <TableHead>Row</TableHead>
-                <TableHead>Student</TableHead>
-                <TableHead>Score</TableHead>
-                <TableHead>Status</TableHead>
-                <TableHead>Issues</TableHead>
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              {preview.rows.map((row) => (
-                <TableRow key={row.rowNumber}>
-                  <TableCell className="align-top font-medium">{row.rowNumber}</TableCell>
-                  <TableCell className="align-top">
-                    <div className="space-y-1">
-                      <p className="font-medium">{row.studentName || "Unnamed student"}</p>
-                      <p className="text-xs text-muted-foreground">{row.studentEmail || "No email"}</p>
-                    </div>
-                  </TableCell>
-                  <TableCell className="align-top">
-                    <div className="space-y-1">
-                      <p className="font-medium">{row.normalizedScore.toFixed(2)}</p>
-                      <p className="text-xs text-muted-foreground">
-                        {Number.isFinite(row.score)
-                          ? `${row.score}/${row.maxScore}`
-                          : row.rubricBreakdown.length > 0
-                            ? "Weighted from rubric columns"
-                            : "No raw score provided"}
-                      </p>
-                    </div>
-                  </TableCell>
-                  <TableCell className="align-top">
-                    <Badge variant={row.accepted ? "outline" : "destructive"}>
-                      {row.accepted ? "Accepted" : "Needs review"}
-                    </Badge>
-                  </TableCell>
-                  <TableCell className="align-top">
-                    {row.issues.length > 0 ? (
-                      <ul className="space-y-1 text-xs text-muted-foreground">
-                        {row.issues.map((issue) => (
-                          <li
-                            key={issue.code}
-                            className={cn(issue.severity === "error" ? "text-destructive" : "text-warning")}
-                          >
-                            {issue.message}
-                          </li>
-                        ))}
-                      </ul>
-                    ) : (
-                      <span className="text-xs text-muted-foreground">No issues</span>
-                    )}
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </div>
-      </div>
-    );
-  };
 
   return (
     <Dialog
@@ -733,76 +441,7 @@ export const HybridGradeImportDialog = ({
               ) : null}
 
               {preview ? (
-                <div className="space-y-3">
-                  <div className="flex flex-wrap items-center gap-2">
-                    <Badge variant="outline">{preview.summary.rowsProcessed} rows scanned</Badge>
-                    <Badge variant="outline" className="border-success/30 bg-success/5 text-success">
-                      {preview.summary.rowsAccepted} accepted
-                    </Badge>
-                    <Badge variant="outline" className="border-warning/30 bg-warning/5 text-warning">
-                      {preview.summary.rowsWithWarnings} with warnings
-                    </Badge>
-                    <Badge variant="outline" className="border-destructive/30 bg-destructive/5 text-destructive">
-                      {preview.summary.rowsRejected} rejected
-                    </Badge>
-                  </div>
-
-                  <div className="overflow-hidden rounded-xl border">
-                    <Table>
-                      <TableHeader>
-                        <TableRow>
-                          <TableHead>Row</TableHead>
-                          <TableHead>Student</TableHead>
-                          <TableHead>Score</TableHead>
-                          <TableHead>Status</TableHead>
-                          <TableHead>Issues</TableHead>
-                        </TableRow>
-                      </TableHeader>
-                      <TableBody>
-                        {preview.rows.map((row) => (
-                          <TableRow key={row.rowNumber}>
-                            <TableCell className="align-top font-medium">{row.rowNumber}</TableCell>
-                            <TableCell className="align-top">
-                              <div className="space-y-1">
-                                <p className="font-medium">{row.studentName || "Unnamed student"}</p>
-                                <p className="text-xs text-muted-foreground">{row.studentEmail || "No email"}</p>
-                              </div>
-                            </TableCell>
-                            <TableCell className="align-top">
-                              <div className="space-y-1">
-                                <p className="font-medium">{row.normalizedScore.toFixed(2)}</p>
-                                <p className="text-xs text-muted-foreground">
-                                  {row.score}/{row.maxScore}
-                                </p>
-                              </div>
-                            </TableCell>
-                            <TableCell className="align-top">
-                              <Badge variant={row.accepted ? "outline" : "destructive"}>
-                                {row.accepted ? "Accepted" : "Needs review"}
-                              </Badge>
-                            </TableCell>
-                            <TableCell className="align-top">
-                              {row.issues.length > 0 ? (
-                                <ul className="space-y-1 text-xs text-muted-foreground">
-                                  {row.issues.map((issue) => (
-                                    <li
-                                      key={issue.code}
-                                      className={cn(issue.severity === "error" ? "text-destructive" : "text-warning")}
-                                    >
-                                      {issue.message}
-                                    </li>
-                                  ))}
-                                </ul>
-                              ) : (
-                                <span className="text-xs text-muted-foreground">No issues</span>
-                              )}
-                            </TableCell>
-                          </TableRow>
-                        ))}
-                      </TableBody>
-                    </Table>
-                  </div>
-                </div>
+                <HybridGradeImportPreview preview={preview} />
               ) : (
                 <div className="rounded-xl border border-dashed bg-muted/20 p-6 text-sm text-muted-foreground">
                   {mode === "csv"
@@ -810,18 +449,6 @@ export const HybridGradeImportDialog = ({
                     : "Upload one or more images, then preview the extracted rows."}
                 </div>
               )}
-
-              {committed ? (
-                <div className="rounded-xl border border-success/30 bg-success/5 p-4">
-                  <div className="flex items-center gap-2 text-success">
-                    <CheckCircle2 className="h-4 w-4" />
-                    <p className="text-sm font-medium">Import completed</p>
-                  </div>
-                  <p className="mt-2 text-xs text-muted-foreground">
-                    {committed.summary.rowsAccepted} rows imported and {committed.summary.createdSyntheticSubmissions} synthetic submissions created.
-                  </p>
-                </div>
-              ) : null}
 
               <div className="flex flex-wrap gap-2">
                 <Button
